# module_V/lab1/openapi-1.2.7.yaml
openapi: 3.0.3
info: 
    title: Swagger for ProductWeb
    version: 1.2.7
    description: Swagger for API by ProductWeb
    contact: 
      email: juliethdiana123@gmail.com
      name: Diana Matamoros 
    license: 
        name: Apache 2.0
        url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
    - url: 'http://localhost:3000/api/v1'
    - url: 'http://172.10.10.1:3000/api/v1'
tags:
    - name: users
      description: Operations about users
    - name: authentication/authorization
      description: Operations about auth & auth
    - name: products
      description: Operations about product  
paths:
  /users:
    get:
      tags: 
        - users
      summary:
        Get all users
      description: Get all users in the platform
      operationId: getAllUsers
      responses:
        '200':
          description: successful operationId
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref : '#/components/schemas/UserDetailResponse'
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Users Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /users/create:
    post:
      tags: 
        - users
      summary: Create users
      description:  This service for create users
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
        required: true
      responses:
        '201':
          description: successful user create 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        '400':
          description: Bad request 
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '404':
          description: Not Found
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/NotFoundError'
  /users/{email}/detail:
    get:
      tags:
        - users
      summary: get user by email
      description: Detail user by email in the platform
      operationId: getUserByEmail
      parameters:
        - name: email
          in: path
          description: The email that need to be fetched.
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: successful operationId
          content:
            application/json:
              schema:
                $ref : '#/components/schemas/UserDetailResponse'
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Users Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /users/{email}/update:
    put:
      tags:
        - users
      summary: Updated user by email
      description: Update User by email. This user can only change phone, first name, address, and last name, in the platform
      operationId: updateUserByEmail
      parameters:
        - name: email
          in: path
          description: The email that need to be fetched.
          required: true
          schema: 
            type: string
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: successful operationId
          content:
            application/json:
              schema:
                $ref : '#/components/schemas/UserDetailResponse'
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Users Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /users/{email}/delete:
    delete:
      tags:
        - users
      summary: Soft Delete user by email
      description: Soft delete User by email.
      operationId: deleteUserByEmail
      parameters:
        - name: email
          in: path
          description: The email that need to be fetched.
          required: true
          schema: 
            type: string 
      responses:
        '204':
          description: No content
          content: { }
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Users Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /auth/login:
    post:
      tags:
        - authentication/authorization
      summary: User authenticate by email and password
      description: Auth users on the platform by email and password, and generate token for session
      operationId: login
      requestBody:
        description: Auth user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuthenticateRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/UserAuthenticateResponse'
        '404':
          description: Users Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /auth/logout:
    post:
      tags:
        - authentication/authorization
      summary: User logout by email and token
      description: Logout users on the platform by email and token, and destroy token for session
      operationId: logOut
      requestBody:
        description: Logout user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogoutRequest'
      responses:
        '204':
          description: No Content
          content: {}
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Users Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /products:
    get:
      tags:
        - product
      summary:
        Get all products 
      description:
        Get all products in active status
      operationId: getAllProduct
      responses:
        '200':
          description: successful operationSku
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref : '#/components/schemas/ProductDetailResponse'
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'          
        '404':
          description: Product Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'              
  /product/create:
    post:
      tags: 
        - product
      summary: Create product
      description:  This service for create product
      operationId: createProduct
      requestBody:
        description: Created product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
        required: true
      responses:
        '201':
          description: successful user create 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCreateResponse'
        '400':
          description: Bad request 
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/BadRequestError' 
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'        
        '404':
          description: Not Found
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/NotFoundError'
  /product/{sku}/detail:
    get:
      tags:
        - product
      summary: Get detail product
      description: Detail product in the platform
      operationId: GetDetailProduct
      parameters:
        - name: sku
          in: path
          description: show selected product detail
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: successful operationSku
          content:
            application/json:
              schema:
                $ref : '#/components/schemas/ProductDetailResponse'
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'        
        '404':
          description: Sku Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /product/{sku}/update:
    put:
      tags:
        - product
      summary: edit product
      description: When updating the product, you will not be able to edit the SKU.
      operationId: updateUserByProduct
      parameters:
        - name: sku
          in: path
          description: The email that need to be fetched.
          required: true
          schema: 
            type: string
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: successful operationId
          content:
            application/json:
              schema:
                $ref : '#/components/schemas/ProductDetailResponse'
        '401':
          description: Invalid Token
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'         
        '404':
          description: Users Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError' 
  /product/{sku}/delete:
    delete:
      tags:
        - product
      summary: Soft Delete user by price
      description: Soft delete User by price.
      operationId: deleteUserByPrice
      parameters:
        - name: sku
          in: path
          description: remove product
          required: true
          schema: 
            type: string 
      responses:
        '200':
          description: Delete 
          content: { }
        '401':
          description: Invalid user
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Users Not Found
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/NotFoundError'               
components:
  schemas:
    UserDetailResponse:
      type: object
      properties:
        id: 
          type: integer
          description: User id on DB
          format: int64
        email:
          type: string
          description: Username for login auth
        first_name:
          type: string
        last_name: 
          type: string
        document_type: 
          type: string
          description: Document type for document ID
        document_id:
          type: string
          description: User document ID
        phone:
          type: string
          description: phone by user
        address:
          type: string
          description: Principal address for User
        status:
          type: number
          description: code for user status
    UserCreateRequest:
      type: object
      properties:
        email:
          type: string
          description: email for user, is unique
        first_name:
          type: string
          description: All names for user
        last_name:
          type: string
          description: Lastname and surname for user
        document_type:
          type: string
          description: Document type for document ID
        document_id:
          type: string
          description: User document ID
        phone:
          type: string
          description: phone by user
        address:
          type: string
          description: Principal address for User
        password:
          type: string
          description: Send by base64 encoded
    UserCreateResponse:
      type: object
      properties: 
        message: 
          type: string
          description: successful create user
    UserUpdateRequest:
      type: object
      properties:
        first_name:
          type: string
          description: All names for user
        last_name:
          type: string
          description: Lastname and surname for user
        phone:
          type: string
          description: phone by user
        address:
          type: string
          description: Principal address for User
    UserAuthenticateRequest:
      type: object
      properties:
        email:
          type: string
          description: email for user, is unique
        password:
          type: string
          description: Send by base64 encoded
    UserAuthenticateResponse:
      type: object
      properties:
        email:
          type: string
          description: email for user, is unique
        id: 
          type: integer
          description: User id on DB
          format: int64
        first_name:
          type: string
        last_name: 
          type: string
        document_type: 
          type: string
          description: Document type for document ID
        document_id:
          type: string
          description: User document ID
        phone:
          type: string
          description: phone by user
        address:
          type: string
          description: Principal address for User
        status:
          type: number
          description: code for user status
        token:
          type: string
          description: JWT for Auth
    UserLogoutRequest:
      type: object
      properties:
        email:
          type: string
          description: email for user, is unique
        date:
          type: string
          description: datetime for logout
    UnauthorizedError:
      type: object
      properties:
        message:
          type: string
          description: Invalid Token
    NotFoundError:
      type: object
      properties:
        code:
          type: number
          description: HTTP Status 404
        message:
          type: string
          description: Not found
    BadRequestError:
      type: object
      properties: 
        error_number:
          type: number
          description: Number of error
        message:
          type: string
          description: message for error description
    ProductDetailResponse:
      type: object
      properties:
        name: 
          type: string
          description: Name of product
        typeOfProduct:
          type: string
          description: Type of product
        price: 
          type: number
          description: Price of product
        sku:
          type: number
          description: Reference number
        provider:
          type: string
          description: Provider product
        coordinates:
          type: object
          description: GEO coordinates of product location
          properties:
            latitude:
              type: number
              format: float
              description: Latitude of the product location
            longitude:
              type: number
              format: float
              description: Longitude of the product location
        ative:
          type: boolean 
    ProductCreateRequest:   
      type: object
      properties:
        name: 
          type: string
          description: Name of product
        typeOfProduct:
          type: string
          description: Type of product
        price: 
          type: number
          description: Price of product
        sku:
          type: number
          description: Reference number
        provider:
          type: string
          description: Provider product
        coordinates:
          type: object
          description: GEO coordinates of product location
          properties:
            latitude:
              type: number
              format: float
              description: Latitude of the product location
            longitude:
              type: number
              format: float
              description: Longitude of the product location
    ProductCreateResponse:
      type: object
      properties: 
        message: 
          type: string
          description: successful create product   
    ProductUpdateRequest:
      type: object
      properties:
        name: 
          type: string
          description: Name of product
        typeOfProduct:
          type: string
          description: Type of product
        price: 
          type: number
          description: Price of product
        provider:
          type: string
          description: Provider product
        coordinates:
          type: object
          description: GEO coordinates of product location
          properties:
            latitude:
              type: number
              format: float
              description: Latitude of the product location
            longitude:
              type: number
              format: float
              description: Longitude of the product location
    